refactor: move GetInstanceIdentity out of ec2 tag
