+++ pkg/util/ec2/ec2_account_id.go
@@ -3,12 +3,11 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-//go:build ec2
-
 package ec2
 
 import (
 	"context"
+	"encoding/json"
 	"fmt"
 
 	pkgconfigsetup "github.com/DataDog/datadog-agent/pkg/config/setup"
@@ -26,4 +25,28 @@ func GetAccountID(ctx context.Context) (string, error) {
 	}
 
 	return ec2id.AccountID, nil
+}
+
+// EC2Identity holds the instances identity document
+// nolint: revive
+type EC2Identity struct {
+	Region     string
+	InstanceID string
+	AccountID  string
+}
+
+// GetInstanceIdentity returns the instance identity document for the current instance
+func GetInstanceIdentity(ctx context.Context) (*EC2Identity, error) {
+	instanceIdentity := &EC2Identity{}
+	res, err := doHTTPRequest(ctx, instanceIdentityURL, useIMDSv2(), true)
+	if err != nil {
+		return instanceIdentity, fmt.Errorf("unable to fetch EC2 API to get identity: %s", err)
+	}
+
+	err = json.Unmarshal([]byte(res), &instanceIdentity)
+	if err != nil {
+		return instanceIdentity, fmt.Errorf("unable to unmarshall json, %s", err)
+	}
+
+	return instanceIdentity, nil
 }
