+++ pkg/util/ec2/ec2_tags.go
@@ -183,30 +183,6 @@ func GetTags(ctx context.Context) ([]string, error) {
 	return tags, err
 }
 
-// EC2Identity holds the instances identity document
-// nolint: revive
-type EC2Identity struct {
-	Region     string
-	InstanceID string
-	AccountID  string
-}
-
-// GetInstanceIdentity returns the instance identity document for the current instance
-func GetInstanceIdentity(ctx context.Context) (*EC2Identity, error) {
-	instanceIdentity := &EC2Identity{}
-	res, err := doHTTPRequest(ctx, instanceIdentityURL, useIMDSv2(), true)
-	if err != nil {
-		return instanceIdentity, fmt.Errorf("unable to fetch EC2 API to get identity: %s", err)
-	}
-
-	err = json.Unmarshal([]byte(res), &instanceIdentity)
-	if err != nil {
-		return instanceIdentity, fmt.Errorf("unable to unmarshall json, %s", err)
-	}
-
-	return instanceIdentity, nil
-}
-
 type ec2SecurityCred struct {
 	AccessKeyID     string
 	SecretAccessKey string
